dodane idDivision do employment

+++ api/ApplicationForAPI.js
@@ -1,8 +1,8 @@
-const ApplicationForRepository = require("../repository/sequelize/ApplicationForRepository");
-const ApplicationForRepositoryMySql2 = require("../repository/mysql2/ApplicationForRepository");
-const StudyRepository = require("../repository/sequelize/StudyRepository");
-const Role = require("../model/Role");
-const Status = require("../model/Status");
+const ApplicationForRepository = require('../repository/sequelize/ApplicationForRepository');
+const ApplicationForRepositoryMySql2 = require('../repository/mysql2/ApplicationForRepository');
+const StudyRepository = require('../repository/sequelize/StudyRepository');
+const Role = require('../model/Role');
+const Status = require('../model/Status');
 
 exports.getApplicationFor = (req, res, next) => {
 	const params = req.query;
@@ -53,7 +53,7 @@ exports.getApplicationForById = (req, res, next) => {
 	ApplicationForRepository.getApplicationForById(appForId).then((appFor) => {
 		if (!appFor) {
 			res.status(404).json({
-				message: "Application for with id: " + appForId + " not found",
+				message: 'Application for with id: ' + appForId + ' not found',
 			});
 		} else {
 			const response = { ...appFor.dataValues, IsStudy: true };
@@ -71,14 +71,14 @@ exports.createApplicationFor = (req, res, next) => {
 			res.status(201).json(newObj);
 		})
 		.catch((err) => {
-			if (err.name === "SequelizeUniqueConstraintError") {
+			if (err.name === 'SequelizeUniqueConstraintError') {
 				res.status(403).json({
-					message: "Użytkownik już złożył wniosek na to szkolenie",
+					message: 'Użytkownik już złożył wniosek na to szkolenie',
 				});
-			} else if (err.name === "SequelizeValidationError") {
-				let message = "";
+			} else if (err.name === 'SequelizeValidationError') {
+				let message = '';
 				for (let m of err.errors) {
-					message += m.message + "\n";
+					message += m.message + '\n';
 				}
 				res.status(403).json({
 					message,
@@ -109,23 +109,23 @@ exports.updateApplicationFor = (req, res, next) => {
 	)
 		.then((result) => {
 			if (result == -1) {
-				res.status(403).json({ message: "Brak uprawnień!" });
+				res.status(403).json({ message: 'Brak uprawnień!' });
 			} else {
 				res.status(200).json({
-					message: "Wniosek zaktualizowany",
+					message: 'Wniosek zaktualizowany',
 					appFor: result,
 				});
 			}
 		})
 		.catch((err) => {
-			if (err.name === "SequelizeUniqueConstraintError") {
+			if (err.name === 'SequelizeUniqueConstraintError') {
 				res.status(403).json({
-					message: "Użytkownik już złożył wniosek o to szkolenie",
+					message: 'Użytkownik już złożył wniosek o to szkolenie',
 				});
-			} else if (err.name === "SequelizeValidationError") {
-				let message = "";
+			} else if (err.name === 'SequelizeValidationError') {
+				let message = '';
 				for (let m of err.errors) {
-					message += m.message + "\n";
+					message += m.message + '\n';
 				}
 				res.status(403).json({
 					message,
@@ -145,32 +145,40 @@ exports.deleteApplicationFor = (req, res, next) => {
 	const appForId = req.params.appForId;
 	const userId = req.userId;
 
-	ApplicationForRepository.deleteApplicationFor(appForId, userId)
-		.then((result) => {
-			if (result == -1) {
-				res.status(403).json({ message: "Brak uprawnień!" });
-			} else {
-				res.status(200).json({
-					message: "Wniosek usunięty",
-					appFor: result,
+	ApplicationForRepository.getApplicationForById(appForId).then((appFor) => {
+		if (appFor.IdPerson != userId) {
+			res
+				.status(403)
+				.json({ message: 'Tylko właściciel wniosku może go usunąć' });
+		} else if (appFor.IdStatus == Status.ZATWIERDZONY_DYR) {
+			res
+				.status(403)
+				.json({ message: 'Nie można usunąć zaakceptowanego wniosku' });
+		} else {
+			ApplicationForRepository.deleteApplicationFor(appForId, userId)
+				.then((result) => {
+					res.status(200).json({
+						message: 'Wniosek usunięty',
+						appFor: result,
+					});
+				})
+				.catch((err) => {
+					if (err.name === 'SequelizeForeignKeyConstraintError') {
+						res.status(403).json({
+							message:
+								'Nie można usunąć wniosku ze względu na przypisane szkolenia',
+						});
+					} else {
+						if (!err.statusCode) {
+							err.statusCode = 500;
+						}
+						res.status(403).json({
+							message: `Nie udało się usunąć wniosku`,
+						});
+					}
 				});
-			}
-		})
-		.catch((err) => {
-			if (err.name === "SequelizeForeignKeyConstraintError") {
-				res.status(403).json({
-					message:
-						"Nie można usunąć wniosku ze względu na przypisane szkolenia",
-				});
-			} else {
-				if (!err.statusCode) {
-					err.statusCode = 500;
-				}
-				res.status(403).json({
-					message: `Nie udało się usunąć wniosku`,
-				});
-			}
-		});
+		}
+	});
 };
 
 exports.getApplicationForByDepId = (req, res, next) => {
@@ -178,8 +186,7 @@ exports.getApplicationForByDepId = (req, res, next) => {
 	ApplicationForRepository.getApplicationForByDepId(depId).then((appFor) => {
 		if (!appFor) {
 			res.status(404).json({
-				message:
-					"Applications with IdDepartment: " + depId + " not found",
+				message: 'Applications with IdDepartment: ' + depId + ' not found',
 			});
 		} else {
 			res.status(200).json(appFor);
@@ -192,8 +199,7 @@ exports.getApplicationForDivId = (req, res, next) => {
 	ApplicationForRepository.getApplicationForByDepId(divId).then((appFor) => {
 		if (!appFor) {
 			res.status(404).json({
-				message:
-					"Applications with IdDivision: " + divId + " not found",
+				message: 'Applications with IdDivision: ' + divId + ' not found',
 			});
 		} else {
 			res.status(200).json(appFor);
@@ -203,16 +209,13 @@ exports.getApplicationForDivId = (req, res, next) => {
 
 exports.getApplicationForStatId = (req, res, next) => {
 	const statId = req.params.statId;
-	ApplicationForRepository.getApplicationForByStatId(statId).then(
-		(appFor) => {
-			if (!appFor) {
-				res.status(404).json({
-					message:
-						"Applications with IdStatus: " + statId + " not found",
-				});
-			} else {
-				res.status(200).json(appFor);
-			}
+	ApplicationForRepository.getApplicationForByStatId(statId).then((appFor) => {
+		if (!appFor) {
+			res.status(404).json({
+				message: 'Applications with IdStatus: ' + statId + ' not found',
+			});
+		} else {
+			res.status(200).json(appFor);
 		}
-	);
+	});
 };
